==> 01_01_sentiments_emoji.ipynb <==
from tqdm import tqdm
import json
import pandas as pd
from pathlib import Path
import numpy as np
import emoji
import re

posts_folder = Path('dataset') / 'emetophobia_posts'
with open(posts_folder / 'emetophobia_all_posts_one_label_normalized.json', 'r') as f:
    all_posts = json.load(f)

len(all_posts)

def minimal_normalziation(text):
    text = text.replace('\n', ' ')
    text = text.replace('\r', ' ')
    return text.strip()

for post in tqdm(all_posts):
    post['content'] = minimal_normalziation(post['content'])

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

analyzer = SentimentIntensityAnalyzer()

for post in tqdm(all_posts):
    post['sentiment'] = analyzer.polarity_scores(post['content'])

def extract_emojis(text):
    return [c for c in text if c in emoji.EMOJI_DATA]

for post in tqdm(all_posts):
    post['emojis'] = extract_emojis(post['content'])
    post['emoji_count'] = emoji.emoji_count(post['content'])

emotions_dir = Path('sentiments')
with open(emotions_dir / 'emetophobia_posts_sent_emoji.json', 'w') as f:
    json.dump(all_posts, f, indent=4)

emoji_dict = {}

for post in tqdm(all_posts):
    for emo in post['emojis']:
        if emo not in emoji_dict:
            emoji_dict[emo] = [post['sentiment']['compound']]
        emoji_dict[emo].append(post['sentiment']['compound'])

len(emoji_dict)

with open(emotions_dir / 'emetophobia_posts_emoji_sentiment.json', 'w') as f:
    json.dump(emoji_dict, f, indent=4)

for k, v in emoji_dict.items():
    emoji_dict[k] = {
        "min": np.min(v),
        "max": np.max(v),
        "mean": np.mean(v),
        "std": np.std(v),
        'nums': len(v)
    }

emoji_dict

df_emoji = pd.DataFrame.from_dict(emoji_dict, orient='index')
df_emoji.head()

df_emoji.to_excel(emotions_dir / 'emetophobia_emojis_statistics.xlsx', index=True, index_label='emoji')

from collections import Counter

all_posts[0]

lbl_counter = Counter()

for post in tqdm(all_posts):
    lbl_counter[post['labels']] += post['emoji_count']

lbl_counter.most_common(20)
