==> 01_03_sentiments_comparison.ipynb <==
import pandas as pd
from pathlib import Path
import json
import numpy as np
from tqdm import tqdm

from scipy.stats import ttest_rel

sentiments_dir = Path('sentiments')

df_emoji = pd.read_excel(sentiments_dir / 'emetophobia_emojis_statistics.xlsx')
df_no_emoji = pd.read_excel(sentiments_dir / 'emetophobia_noemojis_statistics.xlsx')

df_emoji.head()

with open(sentiments_dir / 'emetophobia_posts_emoji_sentiment.json', 'r') as f:
    emoji_dict = json.load(f)
with open(sentiments_dir / 'emetophobia_posts_noemoji_sentiment.json', 'r') as f:
    noemoji_dict = json.load(f)

df_diff = df_emoji.merge(df_no_emoji, on='emoji', how='inner', suffixes=('_with_emoji', '_without_emoji'))

df_diff['mean_difference'] = df_diff['mean_with_emoji'] - df_diff['mean_without_emoji']

# t_stat, p_value = ttest_rel(df_diff['mean_with_emoji'], df_diff['mean_without_emoji'])
for i in tqdm(range(len(df_diff))):
    emoji = df_diff.iloc[i]['emoji']
    if emoji in emoji_dict and emoji in noemoji_dict:
        t_stat, p_value = ttest_rel(emoji_dict[emoji], noemoji_dict[emoji])
        df_diff.at[i, 't_stat'] = t_stat
        df_diff.at[i, 'p_value'] = p_value
    else:
        df_diff.at[i, 't_stat'] = np.nan
        df_diff.at[i, 'p_value'] = np.nan
df_diff = df_diff.sort_values(by='nums_with_emoji', ascending=False)


df_diff.head()

df_diff.to_excel(sentiments_dir / 'emetophobia_diff_statistics.xlsx', index=False)

all_sents_emoji = []
for emoji in emoji_dict:
    all_sents_emoji.extend(emoji_dict[emoji])

all_sents_noemoji = []
for emoji in noemoji_dict:
    all_sents_noemoji.extend(noemoji_dict[emoji])

t_stat, p_value = ttest_rel(all_sents_emoji, all_sents_noemoji)

t_stat, p_value

with open(sentiments_dir / 'emetophobia_posts_sent_emoji.json', 'r') as f:
    posts_emoji = json.load(f)

with open(sentiments_dir / 'emetophobia_posts_sent_noemoji.json', 'r') as f:
    posts_noemoji = json.load(f)

posts_emoji[0], posts_noemoji[0]

all_sents_emoji = []
lbl_sents_emoji = {}

all_sents_noemoji = []
lbl_sents_noemoji = {}

for post in tqdm(posts_emoji):
    all_sents_emoji.append(post['sentiment']['compound'])

    if post['labels'] in lbl_sents_emoji:
        lbl_sents_emoji[post['labels']].append(post['sentiment']['compound'])
    else:
        lbl_sents_emoji[post['labels']] = [post['sentiment']['compound']]

for post in tqdm(posts_noemoji):
    all_sents_noemoji.append(post['sentiment']['compound'])

    if post['labels'] in lbl_sents_noemoji:
        lbl_sents_noemoji[post['labels']].append(post['sentiment']['compound'])
    else:
        lbl_sents_noemoji[post['labels']] = [post['sentiment']['compound']]

all_sents = {
    'max_with_emoji': np.max(all_sents_emoji),
    'min_with_emoji': np.min(all_sents_emoji),
    'mean_with_emoji': np.mean(all_sents_emoji),
    'std_with_emoji': np.std(all_sents_emoji),
    'max_without_emoji': np.max(all_sents_noemoji),
    'min_without_emoji': np.min(all_sents_noemoji),
    'mean_without_emoji': np.mean(all_sents_noemoji),
    'std_without_emoji': np.std(all_sents_noemoji),
}


t_stat_all, p_value_all = ttest_rel(all_sents_emoji, all_sents_noemoji)

all_sents['t_value'] = t_stat_all
all_sents['p_value'] = p_value_all

df_all_sents = pd.DataFrame.from_dict(all_sents, orient='index', columns=['value'])

t_stat_all, p_value_all

df_all_sents.to_excel(sentiments_dir / 'emetophobia_all_sent_statistics.xlsx', index=True, index_label='metric')

lbl_sents = []
for lbl in lbl_sents_emoji:
    if len(lbl_sents_emoji[lbl]) <= 1:
        continue
    
    if lbl in lbl_sents_noemoji:
        t_stat, p_value = ttest_rel(lbl_sents_emoji[lbl], lbl_sents_noemoji[lbl])
    else:
        t_stat, p_value = np.nan, np.nan

    lbl_sents.append({
        'label': lbl,
        'mean_with_emoji': np.mean(lbl_sents_emoji[lbl]),
        'std_with_emoji': np.std(lbl_sents_emoji[lbl]),
        'mean_without_emoji': np.mean(lbl_sents_noemoji[lbl]),
        'std_without_emoji': np.std(lbl_sents_noemoji[lbl]),
        't_stat': t_stat,
        'p_value': p_value
    })

df_lbl_sents = pd.DataFrame.from_records(lbl_sents)

df_lbl_sents.to_excel(sentiments_dir / 'emetophobia_label_sent_statistics.xlsx', index=False)
