==> 02_01_NER_scispacy.ipynb <==
from tqdm import tqdm
import json
import pandas as pd
from pathlib import Path
import numpy as np

import spacy
import scispacy
from scispacy.umls_linking import UmlsEntityLinker
import emoji

posts_folder = Path('dataset') / 'emetophobia_posts'
with open(posts_folder / 'emetophobia_all_posts_one_label_normalized.json', 'r') as f:
    all_posts = json.load(f)

len(all_posts)

ner_dir = Path('ner')

nlp = spacy.load("en_core_sci_md")

nlp.add_pipe("scispacy_linker", config={"resolve_abbreviations": True, "linker_name": "umls"})
linker = UmlsEntityLinker(resolve_abbreviations=True, filter_for_definitions=False)

def minimal_normalziation_noemoji(text):
    text = text.replace('\n', ' ')
    text = text.replace('\r', ' ')
    text = emoji.replace_emoji(text, replace='')
    return text.strip()

for post in tqdm(all_posts):
    doc = nlp(minimal_normalziation_noemoji(post['content']))
    post['entities'] = [(ent, ent._.kb_ents[0][0]) for ent in doc.ents if len(ent._.kb_ents) > 0 and ent._.kb_ents[0][0] in linker.kb.cui_to_entity]

## Top entities

from collections import Counter

counter = Counter()
for post in tqdm(all_posts):
    for ent, cui in post['entities']:
        counter[cui] += 1

counter.most_common(20)

for cui, count in counter.most_common(20):
    print(f"{linker.kb.cui_to_entity[cui]}=====")

## Top entities per label

labels = set()
for post in tqdm(all_posts):
    labels.update([post['labels']])
labels

for label in labels:
    print(label)
    lbl_counter = Counter()
    for post in all_posts:
        if label in post['labels']:
            for ent, cui in post['entities']:
                lbl_counter[cui] += 1
    print(lbl_counter.most_common(5))

    for cui, count in lbl_counter.most_common(5):
        print(f"{linker.kb.cui_to_entity[cui]}=====")
    print('\n\n')

## Total entities per label

for label in labels:
    lbl_counter = 0
    for post in all_posts:
        if label in post['labels']:
            lbl_counter += len(post['entities'])
    
    print(f"{label}: {lbl_counter} entities")

## Statistics entities per post

all_num_entities = []

for post in tqdm(all_posts):
    all_num_entities.append(len(post['entities']))

all_num_entities = np.array(all_num_entities)
print(f"Mean number of entities per post: {np.mean(all_num_entities)}")
print(f"Median number of entities per post: {np.median(all_num_entities)}")
print(f"Max number of entities per post: {np.max(all_num_entities)}")
print(f"Min number of entities per post: {np.min(all_num_entities)}")
print(f"Stdev number of entities per post: {np.std(all_num_entities)}")

## Statistics entities per label

lbl_dict = {}

for label in labels:
    print(label)
    lbl_num_entities = []
    for post in all_posts:
        if label in post['labels']:
            lbl_num_entities.append(len(post['entities']))
    lbl_num_entities = np.array(lbl_num_entities)

    lbl_dict[label] = {
        'mean': np.mean(lbl_num_entities),
        'median': np.median(lbl_num_entities),
        'max': np.max(lbl_num_entities),
        'min': np.min(lbl_num_entities),
        'stdev': np.std(lbl_num_entities),
        'num_posts': len(lbl_num_entities)
    }

    print(f"Mean number of entities per post: {np.mean(lbl_num_entities)}")
    print(f"Median number of entities per post: {np.median(lbl_num_entities)}")
    print(f"Max number of entities per post: {np.max(lbl_num_entities)}")
    print(f"Min number of entities per post: {np.min(lbl_num_entities)}")
    print(f"Stdev number of entities per post: {np.std(lbl_num_entities)}")
    print("\n\n")

df_lbl_scispacy = pd.DataFrame.from_dict(lbl_dict, orient='index')

df_lbl_scispacy.to_excel(ner_dir / 'emetophobia_labels_ner_scispacy.xlsx', index=True, index_label='label')

all_serializable_posts = []

for post in tqdm(all_posts):
    new_post = post.copy()
    new_ents = []
    for ent in new_post['entities']:
        new_ents.append([ent[0].text, ent[1]])
    new_post['entities'] = new_ents
    all_serializable_posts.append(new_post)

len(all_serializable_posts)

with open(ner_dir / 'emetophobia_posts_ner_scispacy.json', 'w') as f:
    json.dump(all_serializable_posts, f, indent=4)

## Posts with 0 entities

posts_0_ents = []

for post in tqdm(all_posts):
    if len(post['entities']) == 0:
        posts_0_ents.append(post['content'])

len(posts_0_ents)

df_posts_0_ents = pd.DataFrame(posts_0_ents, columns=['content'])
df_posts_0_ents.to_excel(ner_dir / 'emetophobia_posts_0_entities_scispacy.xlsx', index=False)
