==> 03_02_DBSCAN_biobert.ipynb <==
from tqdm import tqdm
import json
import pandas as pd
from pathlib import Path
import numpy as np

from sklearn.cluster import KMeans
from gensim.models import Word2Vec

split_dir = Path('dataset') / 'biobert_split'

X_train = np.load(split_dir / 'X_train.npy')
X_train_sent = np.load(split_dir / 'X_train_sent.npy')
X_test = np.load(split_dir / 'X_test.npy')
X_test_sent = np.load(split_dir / 'X_test_sent.npy')
y_train = np.load(split_dir / 'y_train.npy')
y_test = np.load(split_dir / 'y_test.npy')

len(X_train), len(X_test), len(y_train), len(y_test), len(X_train_sent), len(X_test_sent)

X_train = [' '.join(row) for row in X_train]
X_test = [' '.join(row) for row in X_test]

X_train[0], X_test[0]

skipgram_model = Word2Vec(sentences=X_train, vector_size=64, window=10, min_count=1, sg=1, workers=-1)

train_vec_size = 64

X_train_final = []
X_test_final = []

for i in tqdm(range(len(X_train))):
    text = X_train[i]
    sentiment = X_train_sent[i]

    tokens = text.split()
    vectors = [skipgram_model.wv[token] for token in tokens if token in skipgram_model.wv]

    if len(vectors) > train_vec_size:
        vectors = vectors[:train_vec_size]
    elif len(vectors) < train_vec_size:
        vectors += [np.zeros(64)] * (train_vec_size - len(vectors))

    combined_vector = np.concatenate((np.mean(vectors, axis=0), np.array([sentiment])))
    print(len(combined_vector), sentiment, len(vectors))

    X_train_final.append(combined_vector)

for i in tqdm(range(len(X_test))):
    text = X_test[i]
    sentiment = X_test_sent[i]

    tokens = text.split()
    vectors = [skipgram_model.wv[token] for token in tokens if token in skipgram_model.wv]

    if len(vectors) > train_vec_size:
        vectors = vectors[:train_vec_size]
    elif len(vectors) < train_vec_size:
        vectors += [np.zeros(64)] * (train_vec_size - len(vectors))

    combined_vector = np.concatenate((np.mean(vectors, axis=0), np.array([sentiment])))

    X_test_final.append(combined_vector)

X_train_final = np.array(X_train_final)
X_train_final = X_train_final.astype(np.float64)
X_test_final = np.array(X_test_final)
X_test_final = X_test_final.astype(np.float64)

# len(X_train_final), len(X_test_final), len(y_train), len(y_test), X_train_final[0]
    

n_clusters = len(['Question', 'Needing support - Panic attack', 'Rant', 'Potentially Triggering', 'Does Anyone Else...?', 'Needing support: Just not feeling good'])

kmeans = KMeans(n_clusters=n_clusters, random_state=42, n_init=10, max_iter=300, verbose=1)

y_train_pred = kmeans.fit_predict(X_train_final)

from sklearn.metrics import silhouette_score
silhouette_train = silhouette_score(X_train_final, y_train_pred)
print(f'Silhouette score for training set: {silhouette_train}')
