==> 03_03_BERTOPIC_scispacy.ipynb <==
from tqdm import tqdm
import json
import pandas as pd
from pathlib import Path
import numpy as np

from bertopic import BERTopic
from sentence_transformers import SentenceTransformer

split_dir = Path('dataset') / 'scispacy_split'

X_train = np.load(split_dir / 'X_train.npy')
X_train_sent = np.load(split_dir / 'X_train_sent.npy')
X_test = np.load(split_dir / 'X_test.npy')
X_test_sent = np.load(split_dir / 'X_test_sent.npy')
y_train = np.load(split_dir / 'y_train.npy')
y_test = np.load(split_dir / 'y_test.npy')

len(X_train), len(X_test), len(y_train), len(y_test), len(X_train_sent), len(X_test_sent)

X_train = [' '.join([w for w in row if w != '<PAD>']) for row in X_train]
X_test = [' '.join([w for w in row if w != '<PAD>']) for row in X_test]

X_train[0], X_test[0]

# sent_transformer = SentenceTransformer('allenai/scibert_scivocab_uncased')
sent_transformer = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')

X_train_sent_final = []
for row in X_train_sent:
    row = float(row)
    if row > 0.5:
        X_train_sent_final.append('positive')
    elif row < -0.5:
        X_train_sent_final.append('negative')
    else:
        X_train_sent_final.append('neutral')

X_test_sent_final = []
for row in X_test_sent:
    row = float(row)
    if row > 0.5:
        X_test_sent_final.append('positive')
    elif row < -0.5:
        X_test_sent_final.append('negative')
    else:
        X_test_sent_final.append('neutral')

X_train_sent_final[:5], X_test_sent_final[:5]

X_train_final = []
for i in tqdm(range(len(X_train))):
    X_train_final.append(X_train[i] + ' ' + X_train_sent_final[i])

X_test_final = []
for i in tqdm(range(len(X_test))):
    X_test_final.append(X_test[i] + ' ' + X_test_sent_final[i])

X_train_final[:5], X_test_final[:5]

n_clusters = len(['Question', 'Needing support - Panic attack', 'Rant', 'Potentially Triggering', 'Does Anyone Else...?', 'Needing support: Just not feeling good'])

topic_model = BERTopic(verbose=True, n_gram_range=(1, 2), min_topic_size=2, nr_topics=n_clusters, embedding_model=sent_transformer, calculate_probabilities=True)
topics, probs = topic_model.fit_transform(X_train_final)



np.unique(topics)

from sklearn.metrics import silhouette_score

mask = np.array(topics) != -1 
topics_masked = np.array(topics)[mask]
X_train_final_masked = np.array(X_train_final)[mask]

embeddings = topic_model.embedding_model.embedding_model.encode(X_train_final_masked, show_progress_bar=True, convert_to_numpy=True)
silhouette = silhouette_score(embeddings, topics_masked)
print(f'Silhouette Score: {silhouette}')
