==> 04_02_chat_with_documents_scispacy_ner.ipynb <==
from tqdm import tqdm
import json
from pathlib import Path

import spacy
import scispacy
from scispacy.umls_linking import UmlsEntityLinker

ner_dir = Path('ner')
split_dir = Path('dataset') / 'scispacy_split'

with open(ner_dir / 'emetophobia_posts_ner_scispacy.json', 'r') as f:
    all_posts_scispacy = json.load(f)
len(all_posts_scispacy)

with open(ner_dir / 'emetophobia_posts_ner_biobert.json', 'r') as f:
    all_posts_biobert = json.load(f)
len(all_posts_biobert)

linker = UmlsEntityLinker(resolve_abbreviations=True, filter_for_definitions=False)

from langchain.schema import Document
from langchain.vectorstores import Chroma
from langchain_community import embeddings
from langchain.llms import Ollama
from langchain.chains import RetrievalQA

docs_no_ner = []
docs_scispacy_ner = []
docs_biobert_ner = []

for i in tqdm(range(len(all_posts_scispacy))):
    scispacy_post = all_posts_scispacy[i]
    biobert_post = all_posts_biobert[i]

    title = scispacy_post['title']
    content = scispacy_post['content']
    label = scispacy_post['labels']
    ents_scispacy = [e for e in scispacy_post['entities'] if e[1] is not None]
    ents_biobert = [e for e in biobert_post['entities'] if e[1] is not None]

    if len(ents_scispacy) == 0 or len(ents_biobert) == 0:
        continue

    for i in range(len(ents_scispacy)):
        ents_scispacy[i] = f"Entity: {ents_scispacy[i][0]} Canonical Name: {linker.kb.cui_to_entity[ents_scispacy[i][1]].canonical_name} Definition: {linker.kb.cui_to_entity[ents_scispacy[i][1]].definition}"
    
    for i in range(len(ents_biobert)):
        ents_biobert[i] = f"Entity: {ents_biobert[i][0]} Group: {ents_biobert[i][0]}"
    

    # docs_no_ner.append(Document(page_content=title + '\n' + content, metadata={'label': label, 'source': 'no_ner'}))
    docs_scispacy_ner.append(Document(page_content=title + '\n' + content + f"\nDictionary\n" + "\n".join(ents_scispacy), metadata={'label': label, 'source': 'scispacy'}))
    # docs_biobert_ner.append(Document(page_content=title + '\n' + content + f"\nDictionary\n" + "\n".join(ents_biobert), metadata={'label': label, 'source': 'biobert'}))

# docs_no_ner[0], docs_scispacy_ner[0], docs_biobert_ner[0]

def process_question(query, docs):
    embedding = embeddings.OllamaEmbeddings(model="nomic-embed-text")
    vectordb = Chroma.from_documents(docs, embedding)

    llm = Ollama(model="mistral-nemo")

    qa_chain = RetrievalQA.from_chain_type(
        llm=llm,
        retriever=vectordb.as_retriever(),
        return_source_documents=True
    )

    return qa_chain(query)


query = "What are the top 5 most common medical or psychological ideas that the documents refer to? Make a numbered list. Respond with just the list, avoid other text."

process_question(query, docs_scispacy_ner)
